Fixing feed headers, building utils
