+++ src/routes/+layout.svelte
@@ -16,7 +16,7 @@
 </script>
 
 <!-- full page div making the whole screen dark mode -->
-<div class="page-content" data-theme={$themeStore}>
+<div class="page-content content-grid" data-theme={$themeStore}>
 	<Navbar />
 	<slot></slot>
 	<Footer />
@@ -24,21 +24,7 @@
 
 <style>
 .page-content {
-	display: flex;
-	flex-direction: column;
-	flex-wrap: wrap;
-	justify-content: flex-start;
-	align-items: center;
-	width: 100%;
-	height: 100%;
-	padding: 0 12rem 0 12rem;
 	background-color: var(--primary-bg-color);
-}
-
-@media only screen and (max-width: 1366px) {
-	.page-content {
-		padding: 0 4rem;
-	}
 }
 
 @media only screen and (max-width: 430px) {
