+++ src/utils/feed.utils.ts
@@ -1,0 +1,5 @@
+	export const buildShortDesc = (description:string, length:number) => {
+		const shortDesc = description.slice(0, length);
+
+		return shortDesc;
+	}
